requirements.txt
aiogram==3.22.0
python-dotenv==1.0.1
requests==2.32.3
qdrant-client==1.15.1
sentence-transformers==5.1.0
numpy==2.3.2
redis==5.2.1
g4f==0.6.1.4
python-docx==1.2.0
apscheduler==3.11.0
python-dateutil==2.9.0.post0